requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py
beautifulsoup4
Pillow
requests
lxml
psutil
faker
# speedtest-cli